# Improve the performance of put_u8 and put_i8