When using `put_u8/put_i8`, it internally wraps the `u8/i8` into `&[u8]`, and then calls `put_slice()`. Wrapping it into &[u8] is an unnecessary operation. So I optimized this step.

Here's the benchmark below：

**Before**
`test put_u8_bytes_mut         ... bench:         333 ns/iter (+/- 70) = 384 MB/s`

**After**
`test put_u8_bytes_mut         ... bench:         271 ns/iter (+/- 8) = 472 MB/s`