@YHM404 2024-01-31
> This skips some of the length checks making sure there is room for the byte. If the `Buf` is full, this would be UB.

thx. it's fixed.